moved msg_delete_trade_route
